Crate string_error

The string-error crate.

This crate provides a simple way to use a string as an error trait object, i.e. Box<std::error::Error>.

If you need more sophisticated error handling, you should consider error-chain, which also provides functionality to create simple errors from Strings.

Functions

into_err
Creates an error trait object for an owned string (String).

new_err
Creates an error trait object for a string (&str).

static_err
Creates an error trait object for a string constant (&'static str).